Get rid of mypyc_hacks
